[#82] 클래스 구조 리팩토링
- Facade 패턴 적용
