+++ src/main/java/io/driver/codrive/modules/user/service/UserService.java
@@ -10,10 +10,8 @@ import org.springframework.transaction.annotation.Transactional;
 import io.driver.codrive.global.exception.AlreadyExistsApplicationException;
 import io.driver.codrive.global.exception.NotFoundApplcationException;
 import io.driver.codrive.global.util.AuthUtils;
-import io.driver.codrive.global.util.CalculateUtils;
 import io.driver.codrive.modules.follow.domain.Follow;
 import io.driver.codrive.modules.language.service.LanguageService;
-import io.driver.codrive.modules.record.domain.RecordRepository;
 import io.driver.codrive.modules.user.domain.User;
 import io.driver.codrive.modules.user.domain.UserRepository;
 import io.driver.codrive.modules.user.model.request.GoalChangeRequest;
@@ -27,7 +25,6 @@ import lombok.RequiredArgsConstructor;
 public class UserService {
 	private final LanguageService languageService;
 	private final UserRepository userRepository;
-	private final RecordRepository recordRepository;
 
 	public User getUserById(Long userId) {
 		return userRepository.findById(userId)
@@ -93,46 +90,6 @@ public class UserService {
 	@Transactional
 	public List<User> getRandomUsersExcludingMeAndFollowings(User user){
 		return userRepository.getRandomUsersExcludingMeAndFollowings(user.getUserId());
-	}
-
-	@Transactional
-	public void updateSuccessRate(User user) {
-		LocalDate pivotDate = LocalDate.now();
-		int solvedDayCountByWeek = recordRepository.getSolvedDaysByWeek(user.getUserId(), pivotDate);
-		System.out.println("solvedDayCountByWeek: " + solvedDayCountByWeek);
-		int successRate = CalculateUtils.calculateSuccessRate(solvedDayCountByWeek);
-		user.changeSuccessRate(successRate);
-	}
-
-	@Transactional
-	public int getThisWeekRecordsCount(User user) {
-		LocalDate pivotDate = LocalDate.now();
-		return recordRepository.getRecordCountByWeek(user.getUserId(), pivotDate);
-	}
-
-	@Transactional
-	public int getTodayRecordCount(User user) {
-        LocalDateTime startOfDay = LocalDate.now().atStartOfDay(); //오늘 00:00:00
-        LocalDateTime endOfDay = LocalDate.now().atTime(23, 59, 59); //오늘 23:59:59
-		return recordRepository.findAllByUserAndCreatedAtBetween(user, startOfDay, endOfDay).size();
-	}
-
-	@Transactional
-	public UserAchievementResponse getAchievement() {
-		User user = getUserById(AuthUtils.getCurrentUserId());
-		int goal = user.getGoal();
-		int todayCount = getTodayRecordCount(user);
-		int successRate = user.getSuccessRate();
-		int weeklyCountDifference = getWeeklyCountDifference(user);
-		return UserAchievementResponse.of(goal, todayCount, successRate, weeklyCountDifference);
-	}
-
-	@Transactional
-	protected int getWeeklyCountDifference(User user) {
-		int weeklyCount = getThisWeekRecordsCount(user);
-		LocalDate pivotDate = LocalDate.now().minusWeeks(1);
-		int lastWeeklyCount = recordRepository.getRecordCountByWeek(user.getUserId(), pivotDate);
-		return weeklyCount - lastWeeklyCount;
 	}
 
 	@Transactional
